requirements.txt
SQLAlchemy==2.0.36
openai
pymongo
google-api-python-client
google-auth
google-auth-oauthlib
pytz
pandas
mysql.connector
google-cloud-firestore
